compiler/encode: improved dedup algorithm
instead of relying on Hash + Eq on the in-memory representation, we now
use the encoded representation as the source of the hash and equality.

this takes a bit of footwork: we implement a custom hash on the encoded
bytes but then use the range/span of those bytes as the key. since
std HashMap::raw_entry_mut() is still unstable, i've opted to use the
hashbrown crate.