EncodeBuf: higher-level encode_pool_with_header() method
this method uses the knowledge of the pool record size to put the
coverage table closer to the start of the table. previously, the encoded
representation would look like this:

    header
    records
    pool
    coverage

now, it looks like:

    header
    records
    coverage
    pool

feaLib seems to do it this way in some cases. seems like it would be
better just in terms of minimising seeking.